cosmetic changes in ‘speed-type-stats-format’
Tabs are replaced with spaces because tabs are rendered according to
settings such as ‘tab-width’, which is 8 by default, but many people set
it to different values, 4 for example. When ‘tab-width’ is 4 or even 2
the whole thing looks ugly. Spaces help present statistics in a uniform
way.

I've also removed a newline character in ‘speed-type-explaining-message’
because we already have a newline character in format string, so it
resulted in two newlines. I think since you don't use two newlines as
delimiter for more distinct parts of user interface (such as actual text
and statistics), there is no reason to use two newlines here.